// src/app/models/realtor/UpdateRealtorUserDetails.ts
export class UpdateRealtorUserDetails {

   email: string;

   userName: string;

   company: string;

   nmls: string;

   phoneNumber: string;

   brokerName: string;

   managingBroker: string;

   brokerEmail: string;

   brokerPhoneNumber: string;

   closerCompany: string;

   closerName: string;

   closerEmail: string;

   closerPhoneNumber: string;
}